+++ default.project.json
@@ -9,7 +9,7 @@
     "tree": {
 		"$path": "out",
 		
-		"include": {
+		"rbxts_include": {
 			"$path": "include",
 
 			"node_modules": {
