+++ src/index.ts
@@ -11,7 +11,8 @@ import Server from "./lib/net/utilities/server";
 import { Process } from "./lib/process/process";
 import Scheduler from "./lib/process/scheduler";
 import { Users } from "./lib/user";
-import { UserType } from "./lib/user/types";
+import { AbstractUser } from "./lib/user/default";
+import { UserType } from "./lib/user/default/types";
 
 export enum Protocol {
 	/** Create/Load Online User Data */
@@ -116,7 +117,7 @@ namespace Tina {
 	 * @param event event name (defined internally specially for Tina)
 	 * @returns an EventListener object.
 	 */
-	export function when<T extends keyof TinaInternalEvents>(event: T): EventListener<[defined]> {
+	export function when<T extends keyof TinaInternalEvents>(event: T): EventListener<[TinaInternalEvents[T]]> {
 		return TinaEvents.addEventListener(event);
 	}
 
@@ -125,7 +126,9 @@ namespace Tina {
 	 *
 	 * Use the methods on Tina's root (such as `Tina.setUserClass`) to actually apply any modifications.
 	 */
-	export namespace Mirror {}
+	export namespace Mirror {
+		export const User = AbstractUser;
+	}
 }
 
 /** Export Tina itself */
@@ -142,8 +145,8 @@ export { Network } from "./lib/net";
 /** Audience Utility */
 export { Audience } from "./lib/audience/audience";
 
-/** User abstract class */
-export { User, Users } from "./lib/user";
+/** Users namespace */
+export { Users } from "./lib/user";
 
 /** Container export */
 export { Container } from "./lib/container";
