+++ src/lib/core/index.ts
@@ -1,14 +1,15 @@
-import Tina, { User } from "../..";
+import Tina from "../..";
 import { EventEmitter } from "../events";
+import { AbstractUser } from "../user/default";
 
 interface CoreEvents {
 	"player:added": (player: never) => void;
 }
 
-class TinaCore extends EventEmitter<CoreEvents> {}
+class TinaCore extends EventEmitter<CoreEvents> { }
 
 const c = new TinaCore();
 
-c.when("player:added").do((player: User) => {});
+c.when("player:added").do((player: AbstractUser) => { });
 
 export default TinaCore;
