+++ src/lib/net/classes/methods/baseEndpointTypes.d.ts
@@ -1,6 +1,6 @@
-import { User } from "../../../user";
+import { UserType } from "../../../user/default/types";
 
 /**
  * Endpoint insertion for the event types, so we do it and users don't be mad at us for not doing it automatically.
  */
-export type ServerEvent<T> = [user: Array<User & unknown>, ...args: [...[value: T]]];
+export type ServerEvent<T> = [user: Array<UserType>, ...args: [...[value: T]]];
