--- src/lib/user/methods.d.ts
@@ -1,10 +1,0 @@
-/**
- * The FriendPage types
- */
-export interface FriendPage {
-	AvatarFinal: boolean;
-	AvatarUri: string;
-	Id: number;
-	Username: string;
-	IsOnline: boolean;
-}
