+++ src/lib/user/offline/index.ts
@@ -1,0 +1,3 @@
+export class OfflineUser {
+	constructor(userId: number) { }
+}
