+++ src/lib/user/offline/types.d.ts
@@ -1,0 +1,7 @@
+import { DefaultUserImplementation, UserType } from "../default/types";
+
+export interface OfflineUserImplementation extends DefaultUserImplementation {
+	release(): void;
+}
+
+export type OfflineUserType = OfflineUserImplementation & UserType;
